+++ test/notifications/centreNotifications.spec.ts
@@ -304,106 +304,6 @@ describe('Le centre de notifications', () => {
     });
   });
 
-  describe("concernant les tâches liées à l'utilisateur", () => {
-    it("utilise le dépôt de données pour récupérer l'utilisateur", async () => {
-      let idRecu;
-      depotDonnees.utilisateur = async (idUtilisateur) => {
-        idRecu = idUtilisateur;
-        return unUtilisateur().avecId(idUtilisateur).construis();
-      };
-
-      await centreDeNotification().toutesNotifications(unUUID('U1'));
-
-      expect(idRecu).toBe(unUUID('U1'));
-    });
-
-    it("reste robuste si l'utilisateur est introuvable", async () => {
-      depotDonnees.utilisateur = async () => undefined;
-
-      const taches = await centreDeNotification().toutesNotifications(
-        unUUID('U1')
-      );
-
-      expect(taches).toBeInstanceOf(Array);
-    });
-
-    it("renvoie un tableau vide si le profil de l'utilisateur est complet", async () => {
-      referentiel = unReferentiel({
-        nouvellesFonctionnalites: [],
-      });
-
-      depotDonnees.utilisateur = async () =>
-        unUtilisateur()
-          .quiSAppelle('Jean Valjean')
-          .quiTravaillePourUneEntiteAvecSiret('12345')
-          .construis();
-
-      const taches = await centreDeNotification().toutesNotifications(
-        unUUID('U1')
-      );
-
-      expect(taches.length).toBe(0);
-    });
-
-    describe("lorsque le SIRET de l'utilisateur est manquant", () => {
-      beforeEach(() => {
-        depotDonnees.utilisateur = async () =>
-          unUtilisateur()
-            .quiSAppelle('Jeanine Valjean')
-            .quiTravaillePourUneEntiteAvecSiret(null)
-            .construis();
-      });
-
-      it('renvoie la notification correspondant au champ non renseigné du profil', async () => {
-        referentiel = unReferentiel({
-          tachesCompletudeProfil: [{ id: 'siret', titre: 'Titre tâche' }],
-        });
-
-        const taches = await centreDeNotification().toutesNotifications(
-          unUUID('U1')
-        );
-
-        expect(taches.length).toBe(1);
-        expect(taches[0].titre).toBe('Titre tâche');
-        expect(taches[0].statutLecture).toBe('nonLue');
-        expect(taches[0].canalDiffusion).toBe('centreNotifications');
-      });
-
-      it("reste robuste si les données d'une tâche sont absentes du référentiel", async () => {
-        referentiel = unReferentiel({
-          tachesCompletudeProfil: [],
-        });
-
-        const taches = await centreDeNotification().toutesNotifications(
-          unUUID('U1')
-        );
-
-        expect(taches.length).toBe(0);
-      });
-    });
-
-    describe("lorsque l'utilisateur vient d'être invité, donc son profil a plein de champs non renseignés", () => {
-      it('renvoie uniquement la notification « globale » de profil à mettre à jour', async () => {
-        depotDonnees.utilisateur = async () =>
-          unUtilisateur().quiNAPasRempliSonProfil().construis();
-        referentiel = unReferentiel({
-          tachesCompletudeProfil: [
-            { id: 'profil', titre: 'Titre tâche' },
-            { id: 'siret', titre: 'Titre tâche' },
-          ],
-        });
-
-        const taches = await centreDeNotification().toutesNotifications(
-          unUUID('U1')
-        );
-
-        expect(taches.length).toBe(1);
-        expect(taches[0].id).toBe('profil');
-        expect(taches[0].canalDiffusion).toBe('centreNotifications');
-      });
-    });
-  });
-
   describe('sur demande de toutes les notifications', () => {
     beforeEach(() => {
       depotDonnees.utilisateur = async () =>
